services/users/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
autopep8==1.7.0
cachetools==4.2.2
certifi==2022.9.24
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
mysqlclient==2.1.1
psycopg2==2.9.4
pycodestyle==2.9.1
PyJWT==2.5.0
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
tornado==6.1
tzdata==2022.5
tzlocal==4.2
